nuplayer: Allow devices to force a flush when surface changes
This fixes some playback problems on the Zenfone2 (running on 5.0
based blobs).

The problems are easily reproduced by viewing a video in Chrome and
then switching to full-screen mode.  The change in surface causes
the log to get flooded with messages like:

 dc_device_post: Buffer ID=14460 not imported by D

until Android crashes.  With this change, it ends up correctly
mapping the surfaces after the flush fixing the problem.

The code that is being #ifdef'ed out was added in M in commit

   commit a81c6229638a4db56752dd77a6610e0f0971e877
   Author: Lajos Molnar <lajos@google.com>
   Date:   Fri Jul 10 19:17:45 2015 -0700

       nuplayer: use codec->setSurface when possible to avoid seeking

       Bug: 22414719
       Change-Id: I0442e12af960f86a0fc090b4a469c62ba638a1a0

and it appears that our blobs simply cannot handle this optimization
and therefore we need a means by which to disable it.

Change-Id: I105d647e724f5b63ef5cb0852481ecd2990c2d5e